data validations added to DTOs and code improvements

+++ src/main/java/fr/op/taskforce/task/TaskService.java
@@ -2,8 +2,6 @@ package fr.op.taskforce.task;
 
 import fr.op.taskforce.task.dto.TaskDTO;
 import fr.op.taskforce.task.dto.TaskResponseDTO;
-import fr.op.taskforce.task.entity.Task;
-import fr.op.taskforce.user.UserMapper;
 import fr.op.taskforce.user.UserRepository;
 import org.springframework.stereotype.Service;
 
@@ -14,13 +12,11 @@ public class TaskService {
     private final TaskRepository taskRepository;
     private final UserRepository userRepository;
     private final TaskMapper taskMapper;
-    private final UserMapper userMapper;
 
-    public TaskService(TaskRepository taskRepository, UserRepository userRepository) {
+    public TaskService(TaskRepository taskRepository, UserRepository userRepository, TaskMapper taskMapper) {
         this.taskRepository = taskRepository;
         this.userRepository = userRepository;
-        this.taskMapper = new TaskMapper();
-        this.userMapper = new UserMapper();
+        this.taskMapper = taskMapper;
     }
 
 
@@ -29,38 +25,42 @@ public class TaskService {
     }
 
     public TaskResponseDTO getTaskById(Integer id) {
-        return taskMapper.taskToTaskResponseDTO(taskRepository.findById(id).orElse(null));
+        return taskMapper.taskToTaskResponseDTO(taskRepository.findById(id).orElseThrow(() -> new RuntimeException("Task not found with id " + id)));
     }
 
     public TaskResponseDTO saveTask(TaskDTO taskDTO) {
-        var assignedUser = userRepository.findById(taskDTO.assignedToId()).orElse(null);
+        var assignedUser = userRepository.findById(taskDTO.assignedToId()).orElseThrow(() -> new RuntimeException("User not found with id " + taskDTO.assignedToId()));
         return taskMapper.taskToTaskResponseDTO(taskRepository.save(taskMapper.taskDTOToTask(taskDTO, assignedUser)));
     }
 
     public TaskResponseDTO updateTask(Integer id, TaskDTO taskDTO) {
-        var existingTask= taskRepository.findById(id).orElse(new Task());
-            existingTask.setName(taskDTO.name());
-            existingTask.setDescription(taskDTO.description());
-            existingTask.setDueDate(taskDTO.dueDate());
-            existingTask.setStatus(taskDTO.status());
-            existingTask.setAssignedTo(userRepository.findById(taskDTO.assignedToId()).orElse(null));
-            return  taskMapper.taskToTaskResponseDTO(taskRepository.save(existingTask));
+        var existingTask = taskRepository.findById(id).orElseThrow(() -> new RuntimeException("Task not found with id " + id));
+        existingTask.setName(taskDTO.name());
+        existingTask.setDescription(taskDTO.description());
+        existingTask.setDueDate(taskDTO.dueDate());
+        existingTask.setStatus(taskDTO.status());
+        existingTask.setAssignedTo(userRepository.findById(taskDTO.assignedToId()).orElseThrow(() -> new RuntimeException("User not found with id " + taskDTO.assignedToId())));
+        return taskMapper.taskToTaskResponseDTO(taskRepository.save(existingTask));
     }
 
-    public TaskResponseDTO patchTask(Integer id,  TaskDTO taskDTO) {
-        var existingTask= taskRepository.findById(id).orElse(null);
+    public TaskResponseDTO patchTask(Integer id, TaskDTO taskDTO) {
+        var existingTask = taskRepository.findById(id).orElseThrow(() -> new RuntimeException("Task not found with id " + id));
         if (existingTask == null) return null;
         if (taskDTO.name() != null) existingTask.setName(taskDTO.name());
         if (taskDTO.description() != null) existingTask.setDescription(taskDTO.description());
         if (taskDTO.dueDate() != null) existingTask.setDueDate(taskDTO.dueDate());
         if (taskDTO.status() != null) existingTask.setStatus(taskDTO.status());
-        if (taskDTO.assignedToId() != null) existingTask.setAssignedTo(userRepository.findById(taskDTO.assignedToId()).orElse(null));
+        if (taskDTO.assignedToId() != null)
+            existingTask.setAssignedTo(userRepository.findById(taskDTO.assignedToId()).orElseThrow(() -> new RuntimeException("User not found with id " + taskDTO.assignedToId())));
 
-        return  taskMapper.taskToTaskResponseDTO(taskRepository.save(existingTask));
+        return taskMapper.taskToTaskResponseDTO(taskRepository.save(existingTask));
     }
 
     public void deleteTask(Integer id) {
-        // if (taskRepository.existsById(id)) {
+        if (!taskRepository.existsById(id)) {
+            throw new RuntimeException("Task not found with id " + id);
+        }
+
         taskRepository.deleteById(id);
     }
 }
